Added new submission action, Resume
