+++ verifier/submission-verifier.js
@@ -1,4 +1,8 @@
 const ERROR = require("../constants/error-constants");
+const { NEW, IN_PROGRESS, SUBMITTED, RELEASED, COMPLETED, ARCHIVED, CANCELED,
+    REJECTED, WITHDRAWN, ACTIONS } = require("../constants/submission-constants");
+const USER_CONSTANTS = require("../crdc-datahub-database-drivers/constants/user-constants");
+const ROLES = USER_CONSTANTS.USER.ROLES;
 const {toPascalCase} = require("../utility/string-util")
 
 function verifySubmissionAction(submissionId, action){ 
@@ -23,8 +27,8 @@ class SubmissionActionVerifier {
         return this.submission;
     }
 
-    isValidAction(actionMaps){
-        let actionMap = actionMaps?.filter((a)=>a.action === toPascalCase(this.action));
+    isValidAction(){
+        let actionMap = submissionActionMap?.filter((a)=>a.action === toPascalCase(this.action));
         if(!actionMap || actionMap.length === 0)
             throw new Error(`${ERROR.VERIFY.INVALID_SUBMISSION_ACTION} ${this.action}!`);
 
@@ -42,6 +46,26 @@ class SubmissionActionVerifier {
         return this.newStatus;
     }
 }
+
+//actions: NEW, IN_PROGRESS, SUBMITTED, RELEASED, COMPLETED, ARCHIVED, RESUME
+const submissionActionMap = [
+    {action:ACTIONS.SUBMIT, fromStatus: [IN_PROGRESS], 
+        roles: [ROLES.SUBMITTER, ROLES.ORG_OWNER, ROLES.CURATOR,ROLES.ADMIN], toStatus:SUBMITTED},
+    {action:ACTIONS.RELEASE, fromStatus: [SUBMITTED], 
+        roles: [ROLES.CURATOR,ROLES.ADMIN], toStatus:RELEASED},
+    {action:ACTIONS.WITHDRAW, fromStatus: [SUBMITTED], 
+        roles: [ROLES.SUBMITTER, ROLES.ORG_OWNER,], toStatus:WITHDRAWN},
+    {action:ACTIONS.REJECT, fromStatus: [SUBMITTED], 
+        roles: [ROLES.CURATOR,ROLES.ADMIN], toStatus:REJECTED},
+    {action:ACTIONS.COMPLETE, fromStatus: [RELEASED], 
+        roles: [ROLES.CURATOR,ROLES.ADMIN], toStatus:COMPLETED},
+    {action:ACTIONS.CANCEL, fromStatus: [NEW,IN_PROGRESS], 
+        roles: [ROLES.SUBMITTER, ROLES.ORG_OWNER, ROLES.CURATOR,ROLES.ADMIN], toStatus:CANCELED},
+    {action:ACTIONS.ARCHIVE, fromStatus: [COMPLETED], 
+        roles: [ROLES.CURATOR,ROLES.ADMIN], toStatus:ARCHIVED},
+    {action:ACTIONS.RESUME, fromStatus: [REJECTED], 
+            roles: [ROLES.SUBMITTER, ROLES.ORG_OWNER], toStatus:IN_PROGRESS},
+];
 
 module.exports = {
     verifySubmissionAction
